+++ packages/build-next-bricks/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.23.12](https://github.com/easyops-cn/next-core/compare/@next-core/build-next-bricks@1.23.11...@next-core/build-next-bricks@1.23.12) (2025-03-27)
+
+
+### Bug Fixes
+
+* support default value for property signatures ([ac16de9](https://github.com/easyops-cn/next-core/commit/ac16de9f3b0573cb63fade3678e2c473106926db))
+
+
+
+
+
 ## [1.23.11](https://github.com/easyops-cn/next-core/compare/@next-core/build-next-bricks@1.23.10...@next-core/build-next-bricks@1.23.11) (2025-03-13)
 
 **Note:** Version bump only for package @next-core/build-next-bricks