+++ packages/easyops-runtime/CHANGELOG.md
@@ -3,6 +3,57 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.14.0](https://github.com/easyops-cn/next-core/compare/@next-core/easyops-runtime@0.13.9...@next-core/easyops-runtime@0.14.0) (2025-05-29)
+
+
+### Features
+
+* **Router:** Add support for blacklist in app runtime data ([1158c96](https://github.com/easyops-cn/next-core/commit/1158c961308537b9cc78b5200cabb9e62e8de64f))
+
+
+
+
+
+## [0.13.9](https://github.com/easyops-cn/next-core/compare/@next-core/easyops-runtime@0.13.8...@next-core/easyops-runtime@0.13.9) (2025-04-30)
+
+**Note:** Version bump only for package @next-core/easyops-runtime
+
+
+
+
+
+## [0.13.8](https://github.com/easyops-cn/next-core/compare/@next-core/easyops-runtime@0.13.7...@next-core/easyops-runtime@0.13.8) (2025-04-22)
+
+**Note:** Version bump only for package @next-core/easyops-runtime
+
+
+
+
+
+## [0.13.7](https://github.com/easyops-cn/next-core/compare/@next-core/easyops-runtime@0.13.6...@next-core/easyops-runtime@0.13.7) (2025-04-17)
+
+**Note:** Version bump only for package @next-core/easyops-runtime
+
+
+
+
+
+## [0.13.6](https://github.com/easyops-cn/next-core/compare/@next-core/easyops-runtime@0.13.5...@next-core/easyops-runtime@0.13.6) (2025-03-27)
+
+**Note:** Version bump only for package @next-core/easyops-runtime
+
+
+
+
+
+## [0.13.5](https://github.com/easyops-cn/next-core/compare/@next-core/easyops-runtime@0.13.4...@next-core/easyops-runtime@0.13.5) (2025-03-24)
+
+**Note:** Version bump only for package @next-core/easyops-runtime
+
+
+
+
+
 ## [0.13.4](https://github.com/easyops-cn/next-core/compare/@next-core/easyops-runtime@0.13.3...@next-core/easyops-runtime@0.13.4) (2025-03-13)
 
 **Note:** Version bump only for package @next-core/easyops-runtime